refactor: 훅 의존성/요청취소 정리, 배럴 도입으로 임포트 단순화
