+++ src/features/map/index.ts
@@ -1,1 +1,2 @@
 export { default as Map } from "./components/Map";
+export * from "./components";
